Refactored quiz code and implemented createquiz
THINGS TO NOTE:
QuizSummary Temporarily takes a quizId, this was for testing purposes.
Second, I've used an older version of quizhompage.jsp because merge
issues caused the file on github to be corrupted with random characters.
I've added new servlets (CreateQuestion) and jsps to get createquiz.jsp
to work properly.
THINGS TO KEEP IN MIND:
Multiple choice is a bit restrictive at the moment: You may want to try
figuring out how to do it dynamically to allow fewer or more choices
(currently set to 4 choices, 1 correct 3 wrong). Also, a method that
would reorder where the correct answer goes would probably be next on
the TODO list.